+++ mp4/hlspack.go
@@ -161,9 +161,10 @@ func mp4PlanSamples(ps *packagingSource, media []*outTrack) (fts []*fragTrack, o
 			}
 			t.sampleEntry = entry
 		}
-		off, hasCTS, totalTS := fillFragTiming(ft.samples, t.frameDurMs, ft.timescale,
+		off, hasCTS, totalTS, ctsShift := fillFragTiming(ft.samples, t.frameDurMs, ft.timescale,
 			audioGridTS(t, ft.timescale))
 		ft.offsetMs, ft.hasCTS, ft.durMediaTS = off, hasCTS, totalTS
+		ft.ctsShiftTS = ctsShift
 		ft.durMovieMs = totalTS
 		if ft.timescale != movieTimescale {
 			ft.durMovieMs = totalTS * int64(movieTimescale) / int64(ft.timescale)